Targets

This section provides information of available targets (atoms/molecules) in the library.

A target interacts with the laser field and release the electron through multi-photon or tunneling processes. Here we list available targets implemented in the library.

Hydrogen-Like Atom

A hydrogen-like atom has a potential of the form

\[V(r) = -\frac{Z}{\sqrt{r^2+a}},\]

where $Z$ is the nuclear charge number; $a$ denotes the soft-core parameter, which is applied to avoid singularity of the potential and can be adjusted to fit the actual ionization potential of the atom (obtained by TDSE).

The hydrogen-like atom is implemented in the library as Targets.HydrogenLikeAtom.

SemiclassicalSFI.Targets.HydrogenLikeAtom — Type

Represents a hydrogen-like atom.

An instance of HydrogenLikeAtom can be initialized via the constructor method:

HydrogenLikeAtom(Ip, Z, soft_core=1.0, name="[NA]")

Example:

julia> t = Targets.HydrogenLikeAtom(Ip=0.5, Z=1.0, soft_core=1.0, name="H")
[HydrogenLikeAtom] Atom H, Ip=0.5, Z=1.0, SoftCore=1.0

Single-Active-Electron (SAE) Atom

The single-active-electron (SAE) atom is an implementation of the empirical atomic SAE model potential proposed by Tong et al. [Tong_2005] The model potential of the SAE atom has the form

\[V(r) = -\frac{Z + a_1 \mathrm{e}^{-b_1 r} + a_2 r \mathrm{e}^{-b_2 r} + a_3 \mathrm{e}^{-b_3 r}}{r},\]

where the $a_i$ and $b_i$ are tunable model potential parameters [note].

The SAE atom is implemented in the library as Targets.SAEAtom.

SemiclassicalSFI.Targets.SAEAtom — Type

Represents an atom under single-active-electron (SAE) approximation.

An instance of SAEAtom can be initialized via the constructor method:

SAEAtom(Ip, Z, a1=0.0, b1=0.0, a2=0.0, b2=0.0, a3=0.0, b3=0.0, name="[NA]")

Example:

julia> t = Targets.SAEAtom(Ip=0.9035698802, Z=1.0, a1=1.230723, b1=0.6620055, a2=-1.325040, b2=1.236224, a3=-0.2307230, b3=0.4804286, name="He")
[SAEAtom] Atom He, Ip=0.9035698802, Z=1.0

Preset Atom Library

The library provides some preset commonly-used atoms or atomic ions for convenience.

HydrogenLikeAtom: H, He⁺, Li²⁺

SAEAtom: He, Ne, Ne⁺, Ne²⁺, Ar, Ar⁺, Ar²⁺, V, Ni, Kr, Kr⁺, Rb, Nb, Pd, Xe, Xe⁺, Ta

These atoms/ions can be obtained by invoking Targets.**Atom() (for neutral atoms) or Targets.**#pAtom() (for positive atomic ions), where ** denotes the symbol of the nucleus and # denotes the positive charge the ion carries.

Example:

julia> t1 = Targets.HAtom()
[HydrogenLikeAtom] Atom H, Ip=0.5, Z=1.0, SoftCore=1.0

julia> t2 = Targets.Xe1pAtom()
[SAEAtom] Atom Xe⁺, Ip=0.7708, Z=2.0

Molecule

Tong_2005X. M. Tong et al., Empirical Formula for Static Field Ionization Rates of Atoms and Molecules by Lasers in the Barrier-Suppression Regime. J. Phys. B: At. Mol. Opt. Phys. 38, 2593–2600. DOI: 10.1088/0953-4075/38/15/001

noteThe symbols of the parameters are different from that in the original article. $a_1, b_1, a_2, b_2, a_3, b_3$ correspond to $a_1, a_2, a_3, a_4, a_5, a_6$ in the original article respectively.