Targets

This section provides information of available targets (atoms/molecules) in the library.

A target interacts with the laser field and release the electron through multi-photon or tunneling processes. Here we list available targets implemented in the Targets module of the library.

SemiclassicalSFI.Targets — Module

The Targets module provides information about the targeting atoms or molecules.

Targets

Hydrogen-Like Atom

Single-Active-Electron (SAE) Atom

Preset Atom Library

Molecule

Initialization, saving and loading

Molecular-SFI Data Preparation

Molecule's Orientation

Hydrogen-Like Atom

A hydrogen-like atom has a potential of the form

\[V(r) = -\frac{Z}{\sqrt{r^2+a}},\]

where $Z$ is the nuclear charge number; $a$ denotes the soft-core parameter, which is applied to avoid singularity of the potential and can be adjusted to fit the actual ionization potential of the atom (obtained by TDSE).

The hydrogen-like atom is implemented in the library as HydrogenLikeAtom.

SemiclassicalSFI.Targets.HydrogenLikeAtom — Type

struct HydrogenLikeAtom <: SAEAtomBase

Represents a hydrogen-like atom.

An instance of HydrogenLikeAtom can be initialized via the constructor method:

HydrogenLikeAtom(Ip, Z, soft_core=1.0, name="[NA]")

Example:

julia> t = Targets.HydrogenLikeAtom(Ip=0.5, Z=1.0, soft_core=1.0, name="H")
[HydrogenLikeAtom] Atom H, Ip=0.5, Z=1.0, SoftCore=1.0

Single-Active-Electron (SAE) Atom

The single-active-electron (SAE) atom is an implementation of the empirical atomic SAE model potential proposed by Tong et al. [Tong_2005] The model potential of the SAE atom has the form

\[V(r) = -\frac{Z + a_1 \mathrm{e}^{-b_1 r} + a_2 r \mathrm{e}^{-b_2 r} + a_3 \mathrm{e}^{-b_3 r}}{r},\]

where the $a_i$ and $b_i$ are tunable model potential parameters [note].

The SAE atom is implemented in the library as SAEAtom.

SemiclassicalSFI.Targets.SAEAtom — Type

struct SAEAtom <: SAEAtomBase

Represents an atom under single-active-electron (SAE) approximation.

An instance of SAEAtom can be initialized via the constructor method:

SAEAtom(Ip, Z, a1=0.0, b1=0.0, a2=0.0, b2=0.0, a3=0.0, b3=0.0, name="[NA]")

Example:

julia> t = Targets.SAEAtom(Ip=0.9035698802, Z=1.0, a1=1.230723, b1=0.6620055, a2=-1.325040, b2=1.236224, a3=-0.2307230, b3=0.4804286, name="He")
[SAEAtom] Atom He, Ip=0.9035698802, Z=1.0

Preset Atom Library

The library provides some preset commonly-used atoms or atomic ions for convenience.

HydrogenLikeAtom : H, He⁺, Li²⁺

SAEAtom : He, Ne, Ne⁺, Ne²⁺, Ar, Ar⁺, Ar²⁺, V, Ni, Kr, Kr⁺, Rb, Nb, Pd, Xe, Xe⁺, Ta

These atoms/ions can be obtained by invoking Targets.**Atom() (for neutral atoms) or Targets.**#pAtom() (for positive atomic ions), where ** denotes the symbol of the nucleus and # denotes the positive charge the ion carries.

Example:

julia> t1 = Targets.HAtom()[HydrogenLikeAtom] Atom H, Ip=0.5, Z=1.0, SoftCore=1.0
julia> t2 = Targets.Xe1pAtom()[SAEAtom] Atom Xe⁺, Ip=0.7708, Z=2.0

Molecule

The Molecule object represents a generic molecule, which is implemented in the library as Molecule. The structure of Molecule is much more complex than that of atoms because the Molecular ADK (MO-ADK) and Weak-Field Asymptotic Theory (WFAT) features for molecular strong-field ionization require a number of coefficients, which are saved to files for convenience.

Initialization, saving and loading

The Molecule object can be initialized either by providing necessary information of the molecule (mainly atoms, coordinates of the atoms and the charge of the molecule) or from external data (stored in the HDF5 format), cf. the documentation of Molecule:

SemiclassicalSFI.Targets.Molecule — Type

struct Molecule <: Target

Represents a generic molecule.

Fresh initialization

A new instance of Molecule can be initialized via the constructor method:

Molecule(atoms, atom_coords, charge::Integer=0, name::String="[NA]", data_path::String="", calc_energy::Bool=false, rot_α=0.0, rot_β=0.0, rot_γ=0.0)

Parameters:

atoms : Atoms in the molecule, stored as a vector of String.

atom_coords : Atoms' coordinates in the molecule (in Angstrom), stored as a N×3 matrix.

charge : Total charge of the molecule (ion) (optional, default 0).

name : Name of the molecule (optional).

data_path : Path to the molecule's data (default empty). Specifying an empty string indicates no saving (but can still be saved later by calling method MolSaveDataAs).

calc_energy : Indicates whether to calculate the energy data of the molecule upon initialization (default false).

rot_α,rot_β,rot_γ : Euler angles (ZYZ convention) specifying the molecule's orientation (optional, default 0).

Example:

The following example creates a new instance of Hydrogen Molecule, and saves the data related to the Molecule to the specified path.

julia> mol = Targets.Molecule(atoms=["H","H"], atom_coords=[0 0 -0.375; 0 0 0.375], charge=0, name="Hydrogen", data_path="./Molecule_Hydrogen.h5")
[ Info: [Molecule] Data saved for molecule Hydrogen at "./Molecule_Hydrogen.h5".
Molecule [Hydrogen]

Initialization from existing data

To initialize an instance of Molecule from existing external data, invoke

Molecule(data_path::String, rot_α=0.0, rot_β=0.0, rot_γ=0.0)

Example:

julia> mol = Targets.Molecule("./Molecule_Hydrogen.h5")
Molecule [Hydrogen]

The Molecule object, after modification, can be manually saved to a HDF5 file via MolSaveDataAs.

SemiclassicalSFI.Targets.MolSaveDataAs — Function

MolSaveDataAs(mol::Molecule, data_path::String)

Saves the data of the Molecule to the data_path (will change the Molecule's inner field data_path).

Molecular-SFI Data Preparation

To use the molecular strong-field ionization theories such as the MO-ADK and WFAT to provide the intitial conditions of the electrons, the structure coefficients of the Molecule have to be calculated beforehand and stored in the object. Cf. the documentation of MolCalcMOADKCoeff! and MolCalcWFATData!.

Evaluation of the structure coefficients depends on the external quantum chemistry packages. The Targets.MolecularCalculators module undertakes the task of communication with the external quantum packages. Currently only the PySCFMolecularCalculator is implemented.

Customized calculation parameters

When invoking MolCalcMOADKCoeff! and MolCalcWFATData! to perform calculation of structure coefficients, customized calculation parameters can be passed to the kwargs of these methods. These parameters would be passed to the constructor method of the MolecularCalculator (e.g., the basis parameter of the PySCFMolecularCalculator), as well as the MolecularCalculators.calcStructFactorData, MolecularCalculators.calcMOADKCoeff methods. Refer to their documentation below for more information.

SemiclassicalSFI.Targets.MolCalcMOADKCoeff! — Function

MolCalcMOADKCoeff!( mol::Molecule,
                    orbitIdx_relHOMO::Integer = 0,
                    MCType::Type = PySCFMolecularCalculator;
                    kwargs...)

Calculates the MOADK coefficients of the Molecule and saves the data.

MCType : Type of MolecularCalculator if the one for this Molecule is not initialized before. Default is PySCFMolecularCalculator if MCType is not specified.

orbitIdx_relHOMO : Index of selected orbit relative to the HOMO (e.g., 0 indicates HOMO, and -1 indicates HOMO-1) (default 0).

kwargs... : Keyword arguments to pass to the MolecularCalculator, e.g. grid_rNum, l_max.

Example:

julia> mol = Molecule(["H","H"], [0 0 -0.375; 0 0 0.375], 0, "Hydrogen")
Molecule [Hydrogen]

julia> MolCalcMOADKCoeff!(mol, basis="6-31g")
[ Info: [PySCFMolecularCalculator] Running calculation of MOADK coefficients... (ionizing orbital 0 relative to HOMO)

SemiclassicalSFI.Targets.MolCalcWFATData! — Function

MolCalcWFATData!(   mol::Molecule,
                    orbitIdx_relHOMO::Integer = 0,
                    MCType::Type = PySCFMolecularCalculator;
                    kwargs...)

Calculates the WFAT data of the Molecule and saves the data.

MCType : Type of MolecularCalculator if the one for this Molecule is not initialized before. Default is PySCFMolecularCalculator if the MCType is not specified.

orbitIdx_relHOMO : Index of selected orbit relative to the HOMO (e.g., 0 indicates HOMO, and -1 indicates HOMO-1) (default 0).

kwargs... : Keyword arguments to pass to the MolecularCalculator and the calcStructFactorData method, e.g. basis, grid_rNum, grid_rMax, sf_lMax, ⋯

Example:

julia> mol = Targets.Molecule(["H","H"], [0 0 -0.375; 0 0 0.375], 0, "Hydrogen")
Molecule [Hydrogen]

julia> Targets.MolCalcWFATData!(mol, orbitIdx_relHOMO=0, basis="6-31g")
[ Info: [PySCFMolecularCalculator] Running molecular calculation...
[ Info: Finished initialization [taking 0.0466409 second(s)].
[ Info: [PySCFMolecularCalculator] Running calculation of structure factor data... (ionizing orbital 0 relative to HOMO)
✓ Calculating the effective potential... (720000 pts)    Time: 0:00:22
Progress: 100%[●●●●●●●●●●●●●●●●●●●●●●●●●] Time: 0:00:22 (31.08 μs/it)
✓ Calculating the integrals... (7986 integrals)          Time: 0:02:48
Progress: 100%[●●●●●●●●●●●●●●●●●●●●●●●●●] Time: 0:02:48 (21.08 ms/it)

SemiclassicalSFI.Targets.MolecularCalculators — Module

Contains interfaces relating to molecular calculation.

SemiclassicalSFI.Targets.MolecularCalculators.PySCFMolecularCalculator — Type

struct PySCFMolecularCalculator <: MolecularCalculatorBase

An interface of molecular calculation using PySCF.

An instance of PySCFMolecularCalculator can initialized via the following constructor method:

PySCFMolecularCalculator(; mol, basis::String="pc-1", kwargs...)

Parameters

mol::Molecule : The Molecule to be calculated.

basis::String : Basis set used for calculation (default pc-1).

SemiclassicalSFI.Targets.MolecularCalculators.calcStructFactorData — Function

calcStructFactorData(;
    mc::PySCFMolecularCalculator,
    orbitIdx_relHOMO::Int = 0,
    grid_rNum::Int  = 200,
    grid_rMax::Real = 10.,
    grid_θNum::Int  = 60,
    grid_ϕNum::Int  = 60,
    sf_nξMax ::Int  = 5,
    sf_mMax  ::Int  = 5,
    sf_lMax  ::Int  = 10,
    kwargs...)

Calculates the data used in structure factor calculation in WFAT of the given molecule.

Returns

(μ, IntData) : Orbital dipole momentum and the IntData which stores the integrals.

Parameters

mc : The molecular calculator.

orbitIdx_relHOMO: Index of selected orbit relative to the HOMO (e.g., 0 indicates HOMO and -1 indicates HOMO-1) (default 0).

grid_rNum : The number of radial grid (default 200).

grid_rMax : The maximum radius of the radial grid (default 10.0).

grid_θNum : The number of angular grid in the $θ$ direction (default 60).

grid_ϕNum : The number of angular grid in the $ϕ$ direction (default 60).

sf_nξMax : The maximum number of $n_ξ$ used in calculation (default 3).

sf_mMax : The maximum number of $|m|$ used in calculation (default 3).

sf_lMax : The maximum angular quantum number $l$ used in calculation (default 6).

SemiclassicalSFI.Targets.MolecularCalculators.calcMOADKCoeff — Function

calcMOADKCoeff(;
    mc::PySCFMolecularCalculator,
    orbitIdx_relHOMO::Int = 0,
    grid_rNum::Int  = 200,
    grid_rReg::Tuple{<:Real,<:Real} = (3,8),
    grid_θNum::Int  = 60,
    grid_ϕNum::Int  = 60,
    l_max::Int      = 6,
    m_max::Int      = 3,
    kwargs...)

Calculates the MOADK coefficients of the given molecule.

Parameters

mc : The molecular calculator.

orbitIdx_relHOMO: Index of selected orbit relative to the HOMO (e.g., 0 indicates HOMO and -1 indicates HOMO-1) (default 0).

grid_rNum : The number of radial grid (default 200).

grid_rReg : The region of radial distance to fit the wavefunction to obtain the coefficients (default (3,8)).

grid_θNum : The number of angular grid in the $θ$ direction (default 60).

grid_ϕNum : The number of angular grid in the $ϕ$ direction (default 60).

l_max : The maximum number of $l$ calculated (default 6).

m_max : The maximum number of $m$ calculated (default 3).

Molecule's Orientation

The molecule's orientation is described by a set of Euler angles ($z-y'-z''$ convention), which defines a rotational transformation from the molecular frame (MF) to the lab frame (LF). This property of Molecule is NOT included in the saved file and thus needs to be specified each time upon initialization of the Molecule object from external files.

The orientation of the molecule can be obtained and set via the MolRotation and SetMolRotation methods.

SemiclassicalSFI.Targets.MolRotation — Function

MolRotation(mol::Molecule)

Gets the Euler angles (ZYZ convention) specifying the molecule's orientation in format (α,β,γ).

SemiclassicalSFI.Targets.SetMolRotation — Function

SetMolRotation(mol::Molecule, α,β,γ)

Sets the Euler angles (ZYZ convention) specifying the molecule's orientation in format (α,β,γ).

SetMolRotation(mol::Molecule, (α,β,γ))

Sets the Euler angles (ZYZ convention) specifying the molecule's orientation in format (α,β,γ).

Tong_2005X. M. Tong et al., Empirical Formula for Static Field Ionization Rates of Atoms and Molecules by Lasers in the Barrier-Suppression Regime. J. Phys. B: At. Mol. Opt. Phys. 38, 2593–2600. DOI: 10.1088/0953-4075/38/15/001

noteThe symbols of the parameters are different from that in the original article. $a_1, b_1, a_2, b_2, a_3, b_3$ correspond to $a_1, a_2, a_3, a_4, a_5, a_6$ in the original article respectively.